+++ extensions/in-memory/negotiation-store-memory/src/main/java/org/eclipse/dataspaceconnector/negotiation/store/memory/InMemoryContractNegotiationStore.java
@@ -45,8 +45,8 @@ import static java.util.stream.Collectors.toList;
  */
 public class InMemoryContractNegotiationStore implements ContractNegotiationStore {
     private final LockManager lockManager = new LockManager(new ReentrantReadWriteLock());
-    private final Map<String, ContractNegotiation> processesById = new HashMap<>();
-    private final Map<String, ContractNegotiation> processesByCorrelationId = new HashMap<>();
+    private final Map<String, ContractNegotiation> negotiationById = new HashMap<>();
+    private final Map<String, ContractNegotiation> negotiationByCorrelationId = new HashMap<>();
     private final Map<String, ContractNegotiation> contractAgreements = new HashMap<>();
     private final Map<Integer, List<ContractNegotiation>> stateCache = new HashMap<>();
     private final QueryResolver<ContractNegotiation> negotiationQueryResolver = new ReflectionBasedQueryResolver<>(ContractNegotiation.class);
@@ -54,12 +54,12 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
 
     @Override
     public ContractNegotiation find(String id) {
-        return lockManager.readLock(() -> processesById.get(id));
+        return lockManager.readLock(() -> negotiationById.get(id));
     }
 
     @Override
     public @Nullable ContractNegotiation findForCorrelationId(String correlationId) {
-        var process = processesByCorrelationId.get(correlationId);
+        var process = negotiationByCorrelationId.get(correlationId);
         var processId = process != null ? process.getId() : null;
         return find(processId);
     }
@@ -76,8 +76,8 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
             negotiation.updateStateTimestamp();
             delete(negotiation.getId());
             ContractNegotiation internalCopy = negotiation.copy();
-            processesById.put(negotiation.getId(), internalCopy);
-            processesByCorrelationId.put(negotiation.getCorrelationId(), internalCopy);
+            negotiationById.put(negotiation.getId(), internalCopy);
+            negotiationByCorrelationId.put(negotiation.getCorrelationId(), internalCopy);
             var agreement = internalCopy.getContractAgreement();
             if (agreement != null) {
                 contractAgreements.put(agreement.getId(), internalCopy);
@@ -90,7 +90,7 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
     @Override
     public void delete(String processId) {
         lockManager.writeLock(() -> {
-            ContractNegotiation process = processesById.remove(processId);
+            ContractNegotiation process = negotiationById.remove(processId);
             if (process != null) {
                 var tempCache = new HashMap<Integer, List<ContractNegotiation>>();
                 stateCache.forEach((key, value) -> {
@@ -99,7 +99,7 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
                 });
                 stateCache.clear();
                 stateCache.putAll(tempCache);
-                processesByCorrelationId.remove(process.getCorrelationId());
+                negotiationByCorrelationId.remove(process.getCorrelationId());
 
                 if (process.getContractAgreement() != null) {
                     contractAgreements.remove(process.getContractAgreement().getId());
@@ -111,7 +111,7 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
 
     @Override
     public @NotNull Stream<ContractNegotiation> queryNegotiations(QuerySpec querySpec) {
-        return lockManager.readLock(() -> negotiationQueryResolver.query(processesById.values().stream(), querySpec));
+        return lockManager.readLock(() -> negotiationQueryResolver.query(negotiationById.values().stream(), querySpec));
     }
 
     @Override
@@ -148,13 +148,15 @@ public class InMemoryContractNegotiationStore implements ContractNegotiationStor
     }
 
     @Override
-    public Policy findPolicyById(String policyId) {
-        throw new UnsupportedOperationException();
+    public Stream<Policy> findPolicyById(String policyId) {
+        return lockManager.readLock(() -> contractAgreements.values().stream().map(n -> n.getContractAgreement().getPolicy())
+                .filter(p -> Objects.equals(p.getUid(), policyId))
+                .distinct());
     }
 
     @NotNull
     private Stream<ContractAgreement> getAgreements() {
-        return processesById.values().stream()
+        return negotiationById.values().stream()
                 .map(ContractNegotiation::getContractAgreement)
                 .filter(Objects::nonNull);
     }