+++ spi/policy-spi/src/main/java/org/eclipse/dataspaceconnector/spi/policy/store/PolicyArchive.java
@@ -16,6 +16,8 @@ package org.eclipse.dataspaceconnector.spi.policy.store;
 
 import org.eclipse.dataspaceconnector.policy.model.Policy;
 
+import java.util.stream.Stream;
+
 /**
  * Resolves {@linkplain org.eclipse.dataspaceconnector.policy.model.Policy} objects, that are part of a contract agreement.
  * Thus, this archive only houses policies from "foreign" EDC instances.
@@ -23,5 +25,8 @@ import org.eclipse.dataspaceconnector.policy.model.Policy;
 
 @FunctionalInterface
 public interface PolicyArchive {
-    Policy findPolicyById(String policyId);
+    /**
+     * Returns a stream of distinct policies for a given ID.
+     */
+    Stream<Policy> findPolicyById(String policyId);
 }